+++ pybamm/__init__.py
@@ -189,6 +189,11 @@ from .meshes.scikit_fem_submeshes import (
 )
 
 #
+# Serialisation
+#
+from .models.base_model import load_model
+
+#
 # Spatial Methods
 #
 from .spatial_methods.spatial_method import SpatialMethod