+++ pybamm/meshes/zero_dimensional_submesh.py
@@ -38,6 +38,23 @@ class SubMesh0D(SubMesh):
         self.coord_sys = None
         self.npts = 1
 
+    @classmethod
+    def _from_json(cls, snippet):
+        instance = cls.__new__(cls)
+
+        instance.nodes = np.array(snippet["spatial_position"])
+        instance.edges = np.array(snippet["spatial_position"])
+        instance.coord_sys = None
+        instance.npts = 1
+
+        return instance
+
     def add_ghost_meshes(self):
         # No ghost meshes to be added to this class
         pass
+
+    def to_json(self):
+        json_dict = {
+            "spatial_position": self.nodes.tolist(),
+        }
+        return json_dict